refactor: refactor synchronizer.py class
